Update RUN
RUN file now automatically creates the IS_CRYSTAL.txt file in the Crystal directory. The file must be removed after executing the code for the crystal model, however. This is necessary so that the code doesn't recognize any other model as the crystal model. The RUN file automatically removes the file after the crystal is done running.